reorganization + readme updates
